Created the Arena class and was modified the Game class
