--- js/createReloadButton.js
@@ -1,14 +1,0 @@
-import { createElement } from './Helpers/index.js';
-
-export const createReloadButton = () => {
-  const reloadWrap = createElement('div', 'reloadWrap');
-  const reloadButton = createElement('button', 'button');
-  reloadButton.innerText = 'Restart';
-
-  reloadButton.addEventListener('click', () => {
-      window.location.reload();
-    });
-
-  reloadWrap.append(reloadButton);
-  return reloadWrap;
-};
